Add slave mode option to SPI.
This is not a very pretty implementation. You can really only use read
and send on the slave spi since most other functions are implemented
using block. Also it doesn't work with chip select at the moment, just
software enabled SSI.

But it has been tested and it does work.